pyproject.toml
[project]
name = "travel-planner-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "folium>=0.20.0",
    "geopy>=2.4.1",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "streamlit-folium>=0.26.1",
]
